Move snapshots into piet directory
This lets us include the fonts in the snapshots in our package,
which fixes problems when packaging crates that depend on piet.
We can work around that but it feels best to just include these
files.
