+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "snapshots"]
-	path = snapshots
+	path = piet/snapshots
 	url = https://github.com/linebender/piet-snapshots.git
