+++ piet/src/samples/picture_14.rs
@@ -15,7 +15,7 @@ pub fn draw<R: RenderContext>(rc: &mut R) -> Result<(), Error> {
     let text = rc.text();
     let font = text
         .load_font(include_bytes!(
-            "../../../snapshots/resources/Inconsolata-variable.ttf"
+            "../../snapshots/resources/Inconsolata-variable.ttf"
         ))
         .unwrap_or(FontFamily::SYSTEM_UI);
